Display keywords and pull complete contribute.json data.
The build process now downloads the contribute.json file for a project
and saves it, and uses this data to provide a description for each
project (the Github description is used as a fallback) and a list
of keywords.

If the list of keywords is too long, webdev.js hides the extras and
shows a link to un-hide them.

Also includes Babel so that we can write ES6 code. Yay!